requirements.txt
yeelight
discord.py
webcolors
